да, коммит

+++ UserProject/src/main/kotlin/api/requests/CrackHashRequest.kt
@@ -1,4 +1,4 @@
 package org.example.api.requests
 
 @JvmRecord
-data class CrackHashRequest(val hash: String, val maxLength: Short)
+data class CrackHashRequest(val hash: String, val maxLength: Int)
